Fixes by Michael Yevtushenko review
