+++ src/main/java/com/cursor/hw_14_spring_data/controller/BookController.java
@@ -1,9 +1,7 @@
 package com.cursor.hw_14_spring_data.controller;
 
-import com.cursor.hw_14_spring_data.model.Author;
 import com.cursor.hw_14_spring_data.model.Book;
 import com.cursor.hw_14_spring_data.service.LibraryServiceImpl;
-import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,19 +10,13 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@AllArgsConstructor
-public class LibraryController {
+public class BookController {
+
+    private final LibraryServiceImpl LibraryServiceImpl;
 
     @Autowired
-    final LibraryServiceImpl LibraryServiceImpl;
-
-    @PostMapping("/addAuthor")
-    public ResponseEntity addAuthor(@RequestBody Author author) {
-        LibraryServiceImpl.addAuthor(author);
-        return ResponseEntity
-                .status(HttpStatus.CREATED)
-                .build();
-
+    public BookController(LibraryServiceImpl LibraryServiceImpl) {
+        this.LibraryServiceImpl = LibraryServiceImpl;
     }
 
     @PostMapping("/addBook")
@@ -52,25 +44,9 @@ public class LibraryController {
                 .build();
     }
 
-    @DeleteMapping("/deleteAuthor/{id}")
-    public ResponseEntity deleteAuthor(@PathVariable("id") long authorId) {
-        LibraryServiceImpl.deleteAuthor(authorId);
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .build();
-    }
-
     @DeleteMapping("/deleteBook/{id}")
     public ResponseEntity deleteBook(@PathVariable("id") long bookId) {
         LibraryServiceImpl.deleteBook(bookId);
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .build();
-    }
-
-    @PatchMapping("/updateAuthor/{id}")
-    public ResponseEntity updateAuthor(@PathVariable("id") int oldAuthorId, @RequestBody Author author) {
-        LibraryServiceImpl.updateAuthor(oldAuthorId, author);
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .build();
